fixup! More universal SVG stream detection
Extract doctype parsing into a separate file;
Add missing lexical-handler setup.
